// src/app/core/models/skill.enum.ts
export enum SkillCategory {
  TECH = "TECH",
  LANGUAGES = "LANGUAGES",
  ART = "ART",
  MUSIC = "MUSIC",
  COOKING = "COOKING",
  ACADEMICS = "ACADEMICS",
  SPORTS = "SPORTS",
  OTHER = "OTHER"
}

export enum SkillType {
  OFFERED = 'OFFER',
  WANTED = 'REQUEST'
}

export enum SkillStatus {
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  DELETED = 'DELETED'
}